Add useAddProductsFromUrl and use its status for cart loading
In preparation to separate the shopping cart management from the main
checkout page so it can be used elsewhere in calypso, we need to be able
to add the products and coupons from the URL to the cart after the cart
has been initially fetched.

This has presented a problem in the past because until these items have
been added (and the cart refreshed), we cannot be sure that checkout is
"ready". This has major implications for other parts of checkout,
notably useRedirectIfCartEmpty, which needs to wait for this process
before deciding if it should redirect.

Here, we add a new hook, useAddProductsFromUrl, which handles the
process of completing this initialization, and which keeps its own state
that it uses to determine if the process has completed. Other components
and hooks can then rely on this state to know if the cart is "ready" for
use.