Delete the local storage before leaving for another test site
Gutenberg saves the editor mode in the local storage. When the whole upgrade suite runs in a single browser instnace (i.e when running locally without magellan) it's was bound to fail, since once it re-visited a site, it'd open the editor in the code view, which was unexpected. By deleted the local storage data for the site, once it's re-visited, it then falls-back to the default visual editor, as expected.

Also, extract the code to start a new post into its own helper function since it's used in two places at the moment.